get real time updates in the graphs, fixed the food showed
