+++ src/charts/CreateMixCharts.js
@@ -1,70 +1,47 @@
 // @ts-nocheck
 import React, { useEffect, useState } from 'react';
 import RadarChart from "./RadarChart";
-import { IonGrid, IonRow, IonTitle } from '@ionic/react';
-import { IonButton } from '@ionic/react';
-import * as moment from "moment";
+import { IonTitle } from '@ionic/react';
 import PieChart from "./PieChart";
 import { IonItem, IonLabel, IonSelect, IonSelectOption } from '@ionic/react';
 import DataFrame from 'dataframe-js';
 
 // To change the csv path so as to get it from firebase/json
 const CreateFoodCharts = (props) => {
-    
-    //const [testValue, setTestValue] = useState(0);
-
-    // To store count data
-    const [joinedData, setJoinedData] = useState([]);
 
     const [showRadar, setShowRadarChart] = React.useState(true);
     const [showPie, setShowPieChart] = React.useState(true);
 
-    // const x_labels = ['Weekdays']
-
-    // Control button colors
-    const [colorWeekdayButton, setColorWeekdayButton] = React.useState("primary");    
-
     // State the loading variable for the graph, necessary to show the first graph
     const [loading, setLoading] = useState(true);        
 
-    const reestructure_sleep = (d,arra) => {
-        return arra.push( { Time_asleep_seconds: d.Time_asleep_seconds,
-                            End: moment(d.End).subtract(1, "days").format('YYYY-MM-DD'),
-                            Sleep_quality: d.Sleep_quality.slice(0, d.Sleep_quality.indexOf('%')),
-                            Weekday: d.Weekday } );
-    };
+    // For Radar Chart
+    // Function to get the average data over the column_to_groupby variable
+    const getMeanCol = (data_all, column_to_groupby, column_to_mean, nameMeanCol) => {
+        const df = new DataFrame(data_all, Object.keys(data_all[0]));
+        const groupedDF = df.groupBy(column_to_groupby).aggregate(group => group.stat.mean(column_to_mean)).rename('aggregation', nameMeanCol).sortBy(column_to_groupby);
+        return groupedDF.filter(row => !(row.get("Weekday") === null)).toDict();
+      };             
 
-    const reestructure_food = (d,arra) => {
-        return arra.push( { Category: d.Category,
-                            Date: moment(d.Date).format('YYYY-MM-DD'),
-                            Hungry_Overate: d.Hungry_Overate * 100/10,
-                            NonFatty_Fatty: d.NonFatty_Fatty * 100/10,
-                            Unwell_Well: d.Unwell_Well * 100/10 });
-    };      
+    const getWeekdayMeanData = (data, groupbyColumn) => {
+        var sleep_mean = getMeanCol(data, groupbyColumn, 'Sleep_quality', 'QualityMean')
+        var amount_mean = getMeanCol(data, groupbyColumn, 'Amount', 'AmountMean')
+        var wellness_mean = getMeanCol(data, groupbyColumn, 'Wellness' ,'WellnessMean')
+        var fatness_mean = getMeanCol(data, groupbyColumn, 'Fatness' ,'FatnessMean')
 
-    const join = (lookupTable, mainTable, lookupKey, mainKey, select) => {
-        var l = lookupTable.length,
-            m = mainTable.length,
-            lookupIndex = [],
-            output = [];
-        for (var i = 0; i < l; i++) { // loop through l items
-            var row = lookupTable[i];
-            lookupIndex[row[lookupKey]] = row; // create an index for lookup table
-        }
-        for (var j = 0; j < m; j++) { // loop through m items
-            var y = mainTable[j];
-            var x = lookupIndex[y[mainKey]]; // get corresponding row from lookupTable
-            output.push(select(y, x)); // select only the columns you need
-        }
-        return output;
-    };    
-
-    // const buttonClickSetGraph = (name) => {
-    //   if(name == x_labels[0]) {
-    //     setShowRadarChart(true)
-    //     setColorWeekdayButton("primary")
-    //   }
-    // }      
+        var result = []
+        const id_weekday = Object.keys(sleep_mean[groupbyColumn])
+        id_weekday.forEach(function(weekday){
+            var dayData = []
+            // dayData.push(weekday)
+            dayData.push(sleep_mean['QualityMean'][weekday]/10)
+            dayData.push(amount_mean['AmountMean'][weekday]/10)
+            dayData.push(wellness_mean['WellnessMean'][weekday]/10)
+            dayData.push(fatness_mean['FatnessMean'][weekday]/10)
+            result.push( dayData );
+        });
+        return result;
+    };     
 
     // For pie chart
     // Function to group by the key string
@@ -126,70 +103,15 @@ const CreateFoodCharts = (props) => {
         var keys = Object.keys(countByWeekday);
         keys.forEach(function(key){
             countByWeekday[key] = getCountFromArray(categories[key]);
-        });      
+        });     
         return countByWeekday;
-    };         
-
-    const getValues = (dictionary ) => { 
-        var count = []
-        if(dictionary !== undefined || dictionary !== null){
-            var keys = Object.keys(dictionary);
-            keys.forEach(function(key){
-                count.push(dictionary[key]);
-            });
-            var total = count.reduce((a, b) => a + b, 0)
-            return count.map(x => x * 100/total);
-        }
-        else{
-            return count
-        }
-    }    
-
+    };            
     
-    const [weekdayCountData, setWeekdayCountData] = useState([]);
-    const [oneDayIndex, setOneDayIndex] = useState([]);
-    const [oneDayData, setOneDayData] = useState([0,0,0,0,0,0,0,0]);
-    const [labelsPie, setLabelsPie] = useState();
-    
-
-    const selectChanged = (e) => {
-        setOneDayIndex(e.detail.value)
-        setOneDayData(weekdayCountData[e.detail.value])
-      }  
-
-    useEffect(() => {
-        if(!props.loading_sleep && !props.loading_food) {
-            const restr_sleep = []
-            const restr_food = []
-            props.sleepData.map( d => reestructure_sleep(d, restr_sleep))
-            props.foodData.map( d => reestructure_food(d, restr_food))
-            var result = join(restr_sleep, restr_food, "End", "Date", function(food, sleep) {
-                return {
-                    Sleep_quality: (sleep !== undefined) ? sleep.Sleep_quality : null,
-                    Time_asleep: (sleep !== undefined) ? sleep.Time_asleep_seconds : null,
-                    Weekday: (sleep !== undefined) ? sleep.Weekday : null,
-                    Category: food.Category,
-                    Date_only: food.Date,
-                    Amount: food.Hungry_Overate,
-                    Fatness: food.NonFatty_Fatty,
-                    Wellness: food.Unwell_Well
-                };
-            });
-            setJoinedData(result)   
-            const count = getCountByWeekday(result, 'Weekday');
-            setWeekdayCountData(count)
-            setOneDayIndex([])
-            setOneDayData([])
-            setLabelsPie(Object.keys(count[0]))
-            setLoading(false)        
-        } else {
-            console.log("no data to join")
-        }
-    }, []);
+    const [index, setIndex] = useState([]);
+    const labelsPie = ["Caffeinated drinks", "Dairy", "Fruits", "Grains", "Protein food", "Snacks", "Soft drinks", "Vegetables"] 
 
     return (
       <div>
-        {loading && <div>Drawing graph...</div>}
         <IonTitle size="small"
                     style={{
                     textAlign: 'center',
@@ -199,7 +121,7 @@ const CreateFoodCharts = (props) => {
         }}>
             Sleep quality and dinner details by weekday
         </IonTitle>
-        {!loading && showRadar && <RadarChart data={joinedData} name="Radar Graph" loading={loading}/>}
+        {showRadar && <RadarChart data={getWeekdayMeanData(props.joinedData, 'Weekday')} name="Radar Graph" />}
         <IonTitle size="small"
                     style={{
                     textAlign: 'center',
@@ -210,7 +132,7 @@ const CreateFoodCharts = (props) => {
         </IonTitle>   
         <IonItem>
             <IonLabel>Choose a weekday</IonLabel>
-            <IonSelect value={oneDayIndex} placeholder="Select One" onIonChange={selectChanged}>
+            <IonSelect value={index} placeholder="Select One" onIonChange={e => setIndex(e.detail.value)}>
                 <IonSelectOption value="0">Monday</IonSelectOption>
                 <IonSelectOption value="1">Tuesday</IonSelectOption>
                 <IonSelectOption value="2">Wednesday</IonSelectOption>
@@ -220,7 +142,7 @@ const CreateFoodCharts = (props) => {
                 <IonSelectOption value="6">Sunday</IonSelectOption>
             </IonSelect>   
         </IonItem>              
-        {!loading && showPie && <PieChart labels={labelsPie} data={getValues(oneDayData)} index={oneDayIndex} name="Pie Graph" loading={loading} />}
+        {showPie && <PieChart labels={labelsPie} data={getCountByWeekday(props.joinedData, 'Weekday')} index={index} name="Pie Graph" />}
       </div>
     );
 }
