+++ src/charts/RadarChart.js
@@ -1,82 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { Radar } from 'react-chartjs-2';
-import DataFrame from 'dataframe-js';
 
-const RadarChart = (props) => {
-
-    // Function to get the average data over the column_to_groupby variable
-    const getMeanCol = (data_all, column_to_groupby, column_to_mean, nameMeanCol) => {
-        const df = new DataFrame(data_all, Object.keys(data_all[0]));
-        const groupedDF = df.groupBy(column_to_groupby).aggregate(group => group.stat.mean(column_to_mean)).rename('aggregation', nameMeanCol).sortBy(column_to_groupby);
-        return groupedDF.filter(row => !(row.get("Weekday") === null)).toDict();
-      };             
-
-    const getWeekdayMeanData = (data, groupbyColumn) => {
-        var sleep_mean = getMeanCol(data, groupbyColumn, 'Sleep_quality', 'QualityMean')
-        var amount_mean = getMeanCol(data, groupbyColumn, 'Amount', 'AmountMean')
-        var wellness_mean = getMeanCol(data, groupbyColumn, 'Wellness' ,'WellnessMean')
-        var fatness_mean = getMeanCol(data, groupbyColumn, 'Fatness' ,'FatnessMean')
-
-        var result = []
-        const id_weekday = Object.keys(sleep_mean[groupbyColumn])
-        id_weekday.forEach(function(weekday){
-            var dayData = []
-            // dayData.push(weekday)
-            dayData.push(sleep_mean['QualityMean'][weekday]/10)
-            dayData.push(amount_mean['AmountMean'][weekday]/10)
-            dayData.push(wellness_mean['WellnessMean'][weekday]/10)
-            dayData.push(fatness_mean['FatnessMean'][weekday]/10)
-            result.push( dayData );
-        });
-        return result;
-    };         
-
-    // Inital state is empty, required for the component, if not an an error is launched
-    const [barData, setBarData] = useState(                
-        {
-            labels: [
-                'Sleep',
-                'Amount',
-                'Wellness',
-                'Fatness'
-            ],
-            datasets: [
-                {
-                    label: 'Mon.',
-                    data: []
-                }, {
-                    label: 'Tue.',
-                    data: []
-                },
-                {
-                    label: 'Wed.',
-                    data: []
-                },
-                {
-                    label: 'Thu.',
-                    data: []
-                },
-                {
-                    label: 'Fri.',
-                    data: []
-                },
-                {
-                    label: 'Sat.',
-                    data: []
-                },
-                {
-                    label: 'Sun.',
-                    data: []
-                }
-            ]
-        }
-    );
-
-    // Set data
-    useEffect(() => {
-        if (!props.loading) { 
-            var dataToShow = getWeekdayMeanData(props.data, 'Weekday');
-            setBarData(
+const RadarChart = (props) => {               
+    return (
+        <div>
+          <Radar 
+            data={
                 {
                     labels: [
                         'Sleep',
@@ -87,7 +16,7 @@ const RadarChart = (props) => {
                     datasets: [{
                         label: 'Mon.',
                         hidden: false,
-                        data: dataToShow[0],
+                        data: props.data[0],
                         fill: true,
                         backgroundColor: 'rgba(255, 99, 132, 0.2)',
                         borderColor: 'rgb(255, 99, 132)',
@@ -98,7 +27,7 @@ const RadarChart = (props) => {
                       }, {
                         label: 'Tue.',
                         hidden: false,
-                        data: dataToShow[1],
+                        data: props.data[1],
                         fill: true,
                         backgroundColor: 'rgba(54, 162, 235, 0.2)',
                         borderColor: 'rgb(54, 162, 235)',
@@ -110,7 +39,7 @@ const RadarChart = (props) => {
                       {
                         label: 'Wed.',
                         hidden: true,
-                        data: dataToShow[2],
+                        data: props.data[2],
                         fill: true,
                         backgroundColor: 'rgba(255, 159, 64, 0.2)',
                         borderColor: 'rgb(255, 159, 64)',
@@ -122,7 +51,7 @@ const RadarChart = (props) => {
                       {
                         label: 'Thu.',
                         hidden: true,
-                        data: dataToShow[3],
+                        data: props.data[3],
                         fill: true,
                         backgroundColor: 'rgba(255, 205, 86, 0.2)',
                         borderColor: 'rgb(255, 205, 86)',
@@ -134,7 +63,7 @@ const RadarChart = (props) => {
                       {
                         label: 'Fri.',
                         hidden: true,
-                        data: dataToShow[4],
+                        data: props.data[4],
                         fill: true,
                         backgroundColor: 'rgba(75, 192, 192, 0.2)',
                         borderColor: 'rgb(75, 192, 192)',
@@ -146,7 +75,7 @@ const RadarChart = (props) => {
                       {
                         label: 'Sat.',
                         hidden: true,
-                        data: dataToShow[5],
+                        data: props.data[5],
                         fill: true,
                         backgroundColor: 'rgba(153, 102, 255, 0.2)',
                         borderColor: 'rgb(153, 102, 255)',
@@ -158,7 +87,7 @@ const RadarChart = (props) => {
                       {
                         label: 'Sun.',
                         hidden: true,
-                        data: dataToShow[6],
+                        data: props.data[6],
                         fill: true,
                         backgroundColor: 'rgba(255, 153, 255, 0.2)',
                         borderColor: 'rgb(255, 153, 255)',
@@ -168,17 +97,8 @@ const RadarChart = (props) => {
                         pointHoverBorderColor: 'rgb(255, 153, 255)'
                       }
                     ]
-                }            
-            )             
-        } else {
-            console.log("no data")
-        }
-    }, []);            
-
-    return (
-        <div>
-          <Radar 
-            data={barData}
+                }                
+            }
             options={
                  {
                     title: {
